refactor(extract): moves the 'parse' wrappers to their respective folders
