+++ src/extract/clear-caches.mjs
@@ -1,6 +1,6 @@
-import toTypescriptAST from "./parse/to-typescript-ast.mjs";
-import toJavascriptAST from "./parse/to-javascript-ast.mjs";
-import toSwcAST from "./parse/to-swc-ast.mjs";
+import toTypescriptAST from "./tsc/parse.mjs";
+import toJavascriptAST from "./acorn/parse.mjs";
+import toSwcAST from "./swc/parse.mjs";
 import { clearCache as externalModuleHelpers_clearCache } from "./resolve/external-module-helpers.mjs";
 import { clearCache as getManifest_clearCache } from "./resolve/get-manifest.mjs";
 import { clearCache as resolveAMD_clearCache } from "./resolve/resolve-amd.mjs";
