+++ src/extract/swc/parse.mjs
