+++ test/extract/acorn/extract-amd-deps.spec.mjs
@@ -1,6 +1,6 @@
 import { deepEqual } from "node:assert/strict";
 import extractAMDDeps from "#extract/acorn/extract-amd-deps.mjs";
-import { getASTFromSource } from "#extract/parse/to-javascript-ast.mjs";
+import { getASTFromSource } from "#extract/acorn/parse.mjs";
 
 const extractAMD = (
   pJavaScriptSource,
