+++ test/extract/acorn/extract-es6-deps.spec.mjs
@@ -1,6 +1,6 @@
 import { deepEqual } from "node:assert/strict";
 import extractES6Deps from "#extract/acorn/extract-es6-deps.mjs";
-import { getASTFromSource } from "#extract/parse/to-javascript-ast.mjs";
+import { getASTFromSource } from "#extract/acorn/parse.mjs";
 
 const extractES6 = (pJavaScriptSource, pDependencies, pExtension = ".js") =>
   extractES6Deps(
