+++ test/extract/swc/extract-with-swc.utl.mjs
@@ -1,5 +1,5 @@
 import extractSwcDependencies from "#extract/swc/extract-swc-deps.mjs";
-import { getASTFromSource } from "#extract/parse/to-swc-ast.mjs";
+import { getASTFromSource } from "#extract/swc/parse.mjs";
 
 export default (pTypesScriptSource, pExoticRequireStrings = []) =>
   extractSwcDependencies(
