+++ test/extract/tsc/extract-typescript.utl.mjs
@@ -1,5 +1,5 @@
-import extractTypescript from "#extract/ast-extractors/tsc/extract-typescript-deps.mjs";
-import { getASTFromSource } from "#extract/parse/to-typescript-ast.mjs";
+import extractTypescript from "#extract/tsc/extract-typescript-deps.mjs";
+import { getASTFromSource } from "#extract/tsc/parse.mjs";
 
 export default (pTypesScriptSource, pExoticRequireStrings = []) =>
   extractTypescript(
